requirements.txt
azure-mgmt-network
azure-identity
requests
sys
getopt
json
time